Optimize Run + Design Except Feature + Various (#70)
## Optimize config/dir creation
As part of the setup, before running experiments, we create all working directories and place the `config.json` in the folder. 
Until now, this relied on pure Ansible. However, for many jobs, this creation becomes a bottleneck. 
Now, there is a custom module that does this more efficiently.

## Optimize result fetching
We replaced the slow fetching of results with an additional custom module.
More efficiency is possible because now `tsp` or `slurm` do not need to process a completing job id one-by-one.
Instead, they can report that a list of job ids finished and can be downloaded.


## New Experiment Design Feature: `except_filters`
We added a basic implementation of #71 that allows filtering out certain combinations.

## New Super ETL Feature: `pipelines` filter 
We add the possibility when running  a super_etl via make to run only a subset of pipelines (see `pipelines="a b"`)


## ETL Extract Optimization
We only flatten results that require flattening, which speeds up the processing.


## ETL Step
We add a new ETL Loader that allows storing a data frame as a pickle.

## Option to save ETL results to Notion
We add a utility function that allows storing the results of a loader to a Notion page.



---------

Co-authored-by: Hidde L <mail.hidde@gmail.com>